Fixed gloabal search issue

+++ emogo/apps/stream/serializers.py
@@ -142,7 +142,8 @@ class StreamSerializer(DynamicFieldsModelSerializer):
                             itertools.repeat(stream, collaborator_list.__len__()))
                    
         if stream.collaborator_list.count() == 1:
-            if  stream.collaborator_list.all()[0].created_by == self.context.get('request').user:
+            if  stream.collaborator_list.all()[0].created_by == self.context.get('request').user and \
+                stream.collaborator_list.all()[0].created_by.phone_number == self.context.get('request').user.username:
                 self.instance.collaborator_list.filter().delete()
         else:
             return collaborators
